# sivusto/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/quiz-attempts": {
      "get": {
        "tags": [
          "quiz_attempts"
        ],
        "summary": "List Attempts",
        "operationId": "list_attempts",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuizAttemptResponse"
                  },
                  "type": "array",
                  "title": "Response List Attempts"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "quiz_attempts"
        ],
        "summary": "Create Attempt",
        "operationId": "create_attempt",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizAttemptCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizAttemptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/quiz-attempts/{attempt_id}": {
      "get": {
        "tags": [
          "quiz_attempts"
        ],
        "summary": "Get Attempt",
        "operationId": "get_attempt",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizAttemptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "quiz_attempts"
        ],
        "summary": "Delete Attempt",
        "operationId": "delete_attempt",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/iala-lights": {
      "get": {
        "tags": [
          "iala_lights"
        ],
        "summary": "List Iala Lights",
        "description": "List and browse IALA lights with optional category and keyword search.",
        "operationId": "list_iala_lights",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category (e.g. Kardinaalimerkit, Lateraalimerkit)",
              "title": "Category"
            },
            "description": "Filter by category (e.g. Kardinaalimerkit, Lateraalimerkit)"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search by name, rhythm, or description",
              "title": "Search"
            },
            "description": "Search by name, rhythm, or description"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IalaLightResponse"
                  },
                  "title": "Response List Iala Lights"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "iala_lights"
        ],
        "summary": "Create Iala Light",
        "description": "Add a new IALA light to the catalog.",
        "operationId": "create_iala_light",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IalaLightCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IalaLightResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/iala-lights/{light_id}": {
      "get": {
        "tags": [
          "iala_lights"
        ],
        "summary": "Get Iala Light",
        "description": "Fetch a single IALA light by its UUID.",
        "operationId": "get_iala_light",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "light_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Light Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IalaLightResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "iala_lights"
        ],
        "summary": "Update Iala Light",
        "description": "Update an existing IALA light entry.",
        "operationId": "update_iala_light",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "light_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Light Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IalaLightCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IalaLightResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "iala_lights"
        ],
        "summary": "Delete Iala Light",
        "description": "Delete an IALA light from the catalog.",
        "operationId": "delete_iala_light",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "light_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Light Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IalaLightCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "category": {
            "type": "string",
            "maxLength": 100,
            "title": "Category"
          },
          "rhythm": {
            "type": "string",
            "maxLength": 100,
            "title": "Rhythm"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "name",
          "category",
          "rhythm",
          "config"
        ],
        "title": "IalaLightCreate"
      },
      "IalaLightResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "category": {
            "type": "string",
            "maxLength": 100,
            "title": "Category"
          },
          "rhythm": {
            "type": "string",
            "maxLength": 100,
            "title": "Rhythm"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "category",
          "rhythm",
          "config",
          "id",
          "created_at"
        ],
        "title": "IalaLightResponse"
      },
...
      "QuizAttemptCreate": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "quiz_id": {
            "type": "string",
            "format": "uuid",
            "title": "Quiz Id"
          },
          "responses": {
            "items": {
              "$ref": "#/components/schemas/QuizResponseCreate"
            },
            "type": "array",
            "title": "Responses",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "quiz_id"
        ],
        "title": "QuizAttemptCreate"
      },
      "QuizAttemptResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "quiz_id": {
            "type": "string",
            "format": "uuid",
            "title": "Quiz Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "responses": {
            "items": {
              "$ref": "#/components/schemas/QuizResponseItemSchema"
            },
            "type": "array",
            "title": "Responses",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "quiz_id",
          "created_at"
        ],
        "title": "QuizAttemptResponse"
      },
...
      "QuizResponseCreate": {
        "properties": {
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          },
          "selected_option_id": {
            "type": "string",
            "format": "uuid",
            "title": "Selected Option Id"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "selected_option_id"
        ],
        "title": "QuizResponseCreate"
      },
      "QuizResponseItemSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "attempt_id": {
            "type": "string",
            "format": "uuid",
            "title": "Attempt Id"
          },
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          },
          "selected_option_id": {
            "type": "string",
            "format": "uuid",
            "title": "Selected Option Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "attempt_id",
          "question_id",
          "selected_option_id",
          "created_at"
        ],
        "title": "QuizResponseItemSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}